Fix stop_cluster script.

+++ integration-tests/stop_cluster.sh
@@ -24,8 +24,16 @@ if [ -n "$DRUID_INTEGRATION_TEST_SKIP_RUN_DOCKER" ] && [ "$DRUID_INTEGRATION_TES
 
 for node in druid-historical druid-historical-for-query-retry-test druid-coordinator druid-overlord druid-router druid-router-permissive-tls druid-router-no-client-auth-tls druid-router-custom-check-tls druid-broker druid-middlemanager druid-indexer druid-zookeeper-kafka druid-metadata-storage druid-it-hadoop;
 do
-  docker stop $node
-  docker rm $node
+  CONTAINER="$(docker ps -aq -f name=${node})"
+
+  if [ ! -z "$CONTAINER" ]
+  then
+    docker stop $node
+    docker rm $node
+  fi
 done
 
-docker network rm druid-it-net
+if [ ! -z "$(docker network ls -q -f name=druid-it-net)" ]
+then
+  docker network rm druid-it-net
+fi
